fixup! ff2c481cbbd956f668ce65d613e5c9035cefc217
MDEV-33856: Fix MemorySanitizer failures and use a consistent data type

+++ sql/log_event_server.cc
@@ -1461,7 +1461,7 @@ Query_log_event::Query_log_event(THD* thd_arg, const char* query_arg,
   */
   exec_time= ((thd->slave_thread) ?
               thd->orig_exec_time :
-              (ulong) (my_time(0) - thd_arg->start_time));
+              (my_time(0) - thd_arg->start_time));
 
   /**
     @todo this means that if we have no catalog, then it is replicated
